Changes to node, oscar and station
