+++ A1icia CLI/src/com/hulles/a1icia/cli/Version.properties
@@ -1,7 +1,7 @@
 ###Generated file - do not modify!
-#Thu, 21 Dec 2017 02:27:03 -0600
+#Tue, 26 Dec 2017 09:17:23 -0600
 Name=A1icia CLI
 Build-Title=Reykjavik
 Build-Version=0003
-Build-Number=0267
-Build-Date=12/21/2017 02\:27
+Build-Number=0268
+Build-Date=12/26/2017 09\:17
