+++ A1icia Station Installer/src/com/hulles/a1icia/config/station/Version.properties
@@ -1,7 +1,7 @@
 ###Generated file - do not modify!
-#Thu, 21 Dec 2017 02:27:03 -0600
+#Tue, 26 Dec 2017 09:17:23 -0600
 Name=A1icia Station Installer
 Build-Title=Reykjavik
 Build-Version=0002
-Build-Number=0019
-Build-Date=12/21/2017 02\:27
+Build-Number=0020
+Build-Date=12/26/2017 09\:17
